Renamed "ScaledArray" to abstract class "FrameBuffer", implemented "SimpleFrameBuffer" for now. Moved Post-rendering step to "FrameBuffer.update"-method. Scaling still has significant impact on performance.
